+++ templates/fitting-details/utilities/fitting-copy-permalink-to-clipboard.php
@@ -1,0 +1,7 @@
+<div class="fitting-copy-to-clipboard copy-permalink-to-clipboard">
+	<ul class="nav nav-pills nav-stacked">
+		<li role="presentation">
+			<span class="btn btn-default btn-copy-permalink-to-clipboard" data-clipboard-action="copy" data-clipboard-text="<?php echo \get_the_permalink(); ?>"><?php echo \__('Copy permalink to clipboard', 'eve-online-fitting-manager') ; ?></span>
+		</li>
+	</ul>
+</div>